Add an old cgroup v2 path
https://android-review.googlesource.com/c/platform/system/core/+/1324649

+++ native/jni/core/daemon.cpp
@@ -254,19 +254,20 @@ done:
     close(client);
 }
 
-static int switch_cgroup(const char *cgroup, int pid) {
+static void switch_cgroup(const char *cgroup, int pid) {
     char buf[32];
     snprintf(buf, sizeof(buf), "%s/cgroup.procs", cgroup);
-    int fd = open(buf, O_WRONLY | O_APPEND | O_CLOEXEC);
+    if (access(buf, F_OK) != 0)
+        return;
+    int fd = xopen(buf, O_WRONLY | O_APPEND | O_CLOEXEC);
     if (fd == -1)
-        return -1;
+        return;
     snprintf(buf, sizeof(buf), "%d\n", pid);
     if (xwrite(fd, buf, strlen(buf)) == -1) {
         close(fd);
-        return -1;
+        return;
     }
     close(fd);
-    return 0;
 }
 
 static void daemon_entry() {
@@ -299,8 +300,9 @@ static void daemon_entry() {
 
     // Escape from cgroup
     int pid = getpid();
-    if (switch_cgroup("/acct", pid) && switch_cgroup("/sys/fs/cgroup", pid))
-        LOGW("Can't switch cgroup\n");
+    switch_cgroup("/acct", pid);
+    switch_cgroup("/dev/cg2_bpf", pid);
+    switch_cgroup("/sys/fs/cgroup", pid);
 
     // Get self stat
     char buf[64];
@@ -348,13 +350,15 @@ static void daemon_entry() {
     // Use isolated devpts if kernel support
     if (access("/dev/pts/ptmx", F_OK) == 0) {
         auto pts = MAGISKTMP + "/" SHELLPTS;
-        xmkdirs(pts.data(), 0755);
-        xmount("devpts", pts.data(), "devpts",
-               MS_NOSUID | MS_NOEXEC, "newinstance");
-        auto ptmx = pts + "/ptmx";
-        if (access(ptmx.data(), F_OK)) {
-            xumount(pts.data());
-            rmdir(pts.data());
+        if (access(pts.data(), F_OK)) {
+            xmkdirs(pts.data(), 0755);
+            xmount("devpts", pts.data(), "devpts",
+                   MS_NOSUID | MS_NOEXEC, "newinstance");
+            auto ptmx = pts + "/ptmx";
+            if (access(ptmx.data(), F_OK)) {
+                xumount(pts.data());
+                rmdir(pts.data());
+            }
         }
     }
